+++ crates/web-client/test/new_transactions.test.ts
@@ -11,7 +11,6 @@ import {
 
 describe("new_mint_transactions tests", () => {
   it("new_mint_transaction completes successfully", async () => {
-    console.log("starting new_mint_transaction test");
     const { faucetId, accountId } = await setupWalletAndFaucet();
     const result = await mintTransaction(accountId, faucetId);
 
@@ -75,13 +74,18 @@ export const sendTransaction = async (): Promise<SendTransactionResult> => {
     );
     let created_notes = mint_transaction_result.created_notes().notes();
     let created_note_ids = created_notes.map((note) => note.id().to_string());
-    await new Promise((r) => setTimeout(r, 20000)); // TODO: Replace this with loop of sync -> check uncommitted transactions -> sleep
-    await client.sync_state();
+    await window.helpers.waitForTransaction(
+      mint_transaction_result.executed_transaction().id().to_hex()
+    );
 
     await client.fetch_and_cache_account_auth_by_pub_key(senderAccount.id());
-    await client.new_consume_transaction(senderAccount.id(), created_note_ids);
-    await new Promise((r) => setTimeout(r, 20000)); // TODO: Replace this with loop of sync -> check uncommitted transactions -> sleep
-    await client.sync_state();
+    const senderConsumeTransactionResult = await client.new_consume_transaction(
+      senderAccount.id(),
+      created_note_ids
+    );
+    await window.helpers.waitForTransaction(
+      senderConsumeTransactionResult.executed_transaction().id().to_hex()
+    );
 
     await client.fetch_and_cache_account_auth_by_pub_key(senderAccount.id());
     let send_transaction_result = await client.new_send_transaction(
@@ -95,16 +99,18 @@ export const sendTransaction = async (): Promise<SendTransactionResult> => {
     let send_created_note_ids = send_created_notes.map((note) =>
       note.id().to_string()
     );
-    await new Promise((r) => setTimeout(r, 20000)); // TODO: Replace this with loop of sync -> check uncommitted transactions -> sleep
-    await client.sync_state();
+    await window.helpers.waitForTransaction(
+      send_transaction_result.executed_transaction().id().to_hex()
+    );
 
     await client.fetch_and_cache_account_auth_by_pub_key(targetAccount.id());
-    await client.new_consume_transaction(
+    const targetConsumeTransactionResult = await client.new_consume_transaction(
       targetAccount.id(),
       send_created_note_ids
     );
-    await new Promise((r) => setTimeout(r, 20000)); // TODO: Replace this with loop of sync -> check uncommitted transactions -> sleep
-    await client.sync_state();
+    await window.helpers.waitForTransaction(
+      targetConsumeTransactionResult.executed_transaction().id().to_hex()
+    );
 
     const changedSenderAccount = await client.get_account(senderAccount.id());
     const changedTargetAccount = await client.get_account(targetAccount.id());
@@ -327,8 +333,9 @@ export const customTransaction = async (
       transaction_request
     );
     await client.submit_transaction(transaction_result);
-    await new Promise((r) => setTimeout(r, 20000));
-    await client.sync_state();
+    await window.helpers.waitForTransaction(
+      transaction_result.executed_transaction().id().to_hex()
+    );
 
     // Just like in the miden test, you can modify this script to get the execution to fail
     // by modifying the assert
@@ -384,8 +391,9 @@ export const customTransaction = async (
       transaction_request_2
     );
     await client.submit_transaction(transaction_result_2);
-    await new Promise((r) => setTimeout(r, 10000));
-    await client.sync_state();
+    await window.helpers.waitForTransaction(
+      transaction_result_2.executed_transaction().id().to_hex()
+    );
   }, asserted_value);
 };